Remove rendering whitespace, issue 12
The profile_handler js file had a unicode character, U+FEFF,
which is ZERO WIDTH NO-BREAK SPACE. It represents UTF-16 and not
UTF-8. This unicode character is depracated and causes the issues
seen where there is annoying whitespace above the script when it
is appended to the body by Rack::Response.

reference: http://en.wikipedia.org/wiki/Byte_order_mark